create osdMetadataFile and new check_for_osd_support
During cloening of a volume the osdMetadata file schould be created
whenever a volume has osdPolicy set.

The volserver calls check_for_osd_support when moving an OSD-volume
to another server. Unlike in 1.4 and 1.6 this routine calls an RPC
in the AFSVOLOSD set rather than in AFSVol to keep changes to volint.xg
small.