Use ih_ops->close when closing newly created file
For years simply close() was used which in the case of HPSS (or fake HPSS)
files would not close the file. I found this bug when splitting big osd volumes
with more than 500 files to get hard links in fake HPSS. The maximum number of open
filedescriptors was exceeded and the "vos split" failed.