Changes to adapt this code to openafs-osd 1.8
The branch 1.8 of my project opernafs-osd contains the changes which would
hav to integrated into openafs-1.8 in order to prepare it for AFSOSD.
If the directory src/afsosd (where to this github project should be cloned) does not
exist a normal openafs-1.8 is built. With src/afsosd an AFSOSD enabled
AFS is built. All AFSOSD code is in separate shared libraries. The client code
in the kernel (libafs.ko) supports only AFSOSD when the afsd is started with
-libafsosd. Also this openafs-osd 1.8 can be put in production without
any OSD volumes and should be fully compatible with openafs 1.8. It has, however,
some additional commands which also with non-osd-volumes are nice to have.
"vos split" allows to split a large volume at a directory (to be specified as
vnode number). "vos traverse" gives a file size statistic of the contents of
a fileserver or a single volume. "fs stat" gives a statistic about the RPCs.